requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
decorator==4.4.2
future==0.18.2
idna==2.9
pycparser==2.20
PySocks==1.7.1
python-dotenv==0.12.0
python-telegram-bot==12.5.1
requests==2.23.0
six==1.14.0
soupsieve==2.0
tornado==6.0.4
urllib3==1.25.8
